delete file data
